rename fix for promethues_metrics_data
s/promethues_metrics_data/prometheus_metrics_data
