+++ lib/prometheus_metrics_data.rb
